feat: increase training epochs to improve model performance and add experimental network for future use
